Explicitely specify the scope of FileSystem::Type enum.

+++ src/util/guihelpers.cpp
@@ -45,7 +45,7 @@ QIcon createFileSystemColor(FileSystem::Type type, quint32 size)
     QPixmap pixmap(size, size);
     QPainter painter(&pixmap);
     painter.setPen(QColor(0, 0, 0));
-    painter.setBrush(Config::fileSystemColorCode(type));
+    painter.setBrush(Config::fileSystemColorCode(static_cast<int>(type)));
     painter.drawRect(QRect(0, 0, pixmap.width() - 1, pixmap.height() - 1));
     painter.end();
